requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
boto3==1.37.38
botocore==1.37.38
brokenaxes==0.6.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.8.1
eval_type_backport==0.2.2
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
ftfy==6.2.3
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
h5py==3.11.0
huggingface-hub==0.30.2
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
nvitop==1.4.2
oauthlib==3.2.2
opencv-contrib-python==4.12.0.88
packaging==25.0
pandas==2.0.3
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
portalocker==3.0.0
prettytable==3.11.0
protobuf==5.29.4
psutil==7.0.0
ptflops==0.7.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pyecharts==2.0.8
pynvml==11.5.3
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.11.5
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentry-sdk==2.27.0
setproctitle==1.3.5
setuptools==75.1.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
swankit==0.2.0
swanlab==0.6.0
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.15
torch==2.1.0
torchaudio==2.1.0
torchinfo==1.8.0
torchvision==0.16.0
tqdm==4.67.1
triton==2.1.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
wandb==0.19.10
wcwidth==0.2.13
Werkzeug==3.0.6
wheel==0.44.0
yacs==0.1.8
zipp==3.20.2
